Move BeaconRolesService to service directory
