+++ src/main/java/uk/gov/mca/beacons/api/hateoas/BeaconLinkStrategy.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 import uk.gov.mca.beacons.api.controllers.BeaconsController;
 import uk.gov.mca.beacons.api.dto.WrapperDTO;
 import uk.gov.mca.beacons.api.jpa.entities.Beacon;
+import uk.gov.mca.beacons.api.services.BeaconRolesService;
 
 @Service
 public class BeaconLinkStrategy implements IHateoasLinkStrategy<Beacon> {
